Unify logic for eraseText and inputText commands on iOS (#1417)
